+++ nsd/exec.c
@@ -221,7 +221,7 @@ Ns_WaitForProcess(int pid, int *exitcodePtr)
 
 int
 Ns_ExecArgblk(char *exec, char *dir, int fdin, int fdout,
-		char *args, Ns_Set *env)
+	      char *args, Ns_Set *env)
 {
 #ifndef _WIN32
     int    pid;
@@ -308,10 +308,10 @@ Ns_ExecArgblk(char *exec, char *dir, int fdin, int fdout,
 
         s = args;
         while (*s != '\0') {
-            int             len;
+            size_t len;
 
             len = strlen(s);
-            Ns_DStringNAppend(&cds, s, len);
+            Ns_DStringNAppend(&cds, s, (int)len);
             s += len + 1;
             if (*s != '\0') {
                 Ns_DStringNAppend(&cds, " ", 1);
@@ -385,7 +385,7 @@ Ns_ExecArgv(char *exec, char *dir, int fdin, int fdout,
     } else {
         int  i;
         for (i = 0; argv[i] != NULL; ++i) {
-            Ns_DStringNAppend(&ads, argv[i], strlen(argv[i]) + 1);
+	  Ns_DStringNAppend(&ads, argv[i], (int)strlen(argv[i]) + 1);
         }
         args = ads.string;
     }
